\name{simTile}

\docType{data}

\alias{simTile}

\title{Simulated tiling array data set containing Loci of Enhanced
  Significance
}

\description{
  This data set consists of simulated tiling array data between 2
  conditions with 3 arrays each. It contains 1000 equally spaced probes
  and 2 regions of regulation covering 50 probes each. The expression
  values were generated by constant fold change in the regulation
  regions and additive as well as multiplicative noise sources. The
  expression values and annotation are stored in an object of class 'eset'.
}

\usage{data(simTile)}

\format{An expression set ('eset')}

\source{Julian Gehring}

\keyword{datasets}
