\name{statTile}

\docType{data}

\alias{statTile}
\alias{borders}
\alias{chr}
\alias{pos}
\alias{pval}

\title{Probe-level statistics from simulated tiling array data set}

\description{
  This data set consists of simulated p-values for a tiling array
  experiments. It contains 1000 equally spaced probes and 2 regions of
  regulation covering 50 probes each.
}

\usage{data(statTile)}

\format{
  Vectors 'pos', 'pval' and 'chr' with probe positions, corresponding
  p-values and chromosomal locations. Vector 'borders' with the boundary
  positions of the regions with simulated effect.
}

\source{Julian Gehring}

\keyword{datasets}
